+++ server/src/voxel/chunk.c
@@ -29,7 +29,7 @@ chunk *chunkNew(uint16_t x,uint16_t y,uint16_t z){
 	c->y = y & (~0xF);
 	c->z = z & (~0xF);
 	c->nextFree = NULL;
-	c->clientsUpdated = 1 << 63;
+	c->clientsUpdated = (uint64_t)1 << 63;
 	
 	memset(c->data,0,sizeof(c->data));
 	return c;
